Rename 1923b.txt to 1981.txt
This text is based on the 1923 version of "Sunday Morning", and it incorporates an imitation work done by Lysander Kemp, published in a literary magazine The Hudson Review in 1981. Kemp's work captures the spirit of how Wallace Stevens and William Carlos Williams--now a prominent American Modernist poet figures--would have peer reviewed their work.